backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==1.10.13
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pymupdf==1.23.3
python-docx==0.8.11
sentence-transformers==2.2.2
numpy==1.25.2
scikit-learn==1.3.0
psycopg2-binary==2.9.7
alembic==1.12.0
pytest==7.4.2
httpx==0.25.0